Fixed banner issue and changed the offline online texts
