+++ public/locales/en/common.json
@@ -1,10 +1,10 @@
 {
   "company": "Cube Design & Communication",
   "company_shorthand": "Cube Design",
-  "printing": "Offline",
-  "digital": "Online",
-  "dynamic_printing": "Offline Solution",
-  "dynamic_digital": "Online Solutions",
+  "printing": "Printing",
+  "digital": "Digital",
+  "dynamic_printing": "Printing Solution",
+  "dynamic_digital": "Digital Solutions",
   "contactus": "Contact Us",
   "consultus": "Get a free consulting!",
   "pagenotfoundheader": "Page Not Found", 
@@ -16,7 +16,7 @@
   "chatbot" : {
     "init1" : "Hi, there! Having a hard time to find what you need? Let me help you!",
     "init2" : "Please select you are interested!",
-    "i1" : "Offline",
+    "i1" : "Printing",
     "p1" : "Printing & Installation",
     "p1a" : "Large Format Printing",
     "p1b" : "Digital Printing Service",
@@ -26,7 +26,7 @@
     "p2a" : "Directional Signage Design",
     "p3" : "Promotional Products",
     "p3a" : "Promotional Product & Merchandise",
-    "i2" : "Online",
+    "i2" : "Digital",
     "d1" : "Website Design & Development",
     "d1a" : "Search Engine Optimization",
     "d1b" : "Content Management System",
@@ -61,9 +61,9 @@
   "home": {
     "banner1": "Cube Design & Communication",
     "banner2": "makes Poster Design",
-    "printing": "Offline",
+    "printing": "Printing",
     "printing_desc": "Maximize your visual impact with our professional printing and installation services, including way finding and promotional products.",
-    "digital": "Online",
+    "digital": "Digital",
     "digital_desc": "Establish your online presence with our full-service solutions tailored to enhance your business.",
     "brand_value": "Brand Value",
     "onestop_solution": "One-Stop Solution",
